MH-130 #time 30m MH-130 #done MH-130 #comment ResponseEntitys und Exceptionhandling hinzugefügt
MH-131 #time 30m
MH-131 #done
MH-131 #comment ResponseEntitys und Exceptionhandling hinzugefügt

+++ backend/src/main/java/de/szut/backend/service/CategoryService.java
@@ -1,16 +1,23 @@
 package de.szut.backend.service;
 
+import de.szut.backend.exceptions.CategoryBoundToExpenseException;
 import de.szut.backend.model.Categorys.Category;
+import de.szut.backend.model.Expenses.Expense;
 import de.szut.backend.repository.CategoryRepository;
+import de.szut.backend.repository.ExpensesRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
+
 import java.util.List;
 
 @Service
 public class CategoryService {
     private CategoryRepository c_Repository;
+    private ExpensesRepository ex_Repository;
 
-    public CategoryService(CategoryRepository c_Repository){
+    public CategoryService(CategoryRepository c_Repository, ExpensesRepository ex_Repository){
         this.c_Repository = c_Repository;
+        this.ex_Repository = ex_Repository;
     }
 
     public Category createCategory(Category categoryToCreate){
@@ -22,7 +29,11 @@ public class CategoryService {
             return this.c_Repository.save(categoryToCreate);
     }
 
-    public void deleteCategoryById(long categoryId){
+    public void deleteCategoryById(long categoryId) throws CategoryBoundToExpenseException {
+        Expense expenseWithCategory = ex_Repository.findFirstByCategoryId(categoryId);
+        if(expenseWithCategory != null){
+            throw new CategoryBoundToExpenseException();
+        }
         this.c_Repository.deleteById(categoryId);
     }
